+++ CHANGELOG.rst
@@ -6,12 +6,18 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.1.0/>`__,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`__.
 
+..
+    2.0.0 suggested for next version per CLI output change
 `Unreleased`_
 -------------
 
 Added
 ~~~~~
 - Add ``RestClient.search`` method for performing package searches using the V3 API
+
+Changed
+~~~~~~~
+- ``dpres-client search`` command now uses V3 API
 
 Fixed
 ~~~~~